+++ src/main/java/org/jabref/logic/integrity/IntegrityCheck.java
@@ -52,6 +52,7 @@ public class IntegrityCheck {
             entryCheckers.addAll(List.of(
                     new ASCIICharacterChecker(),
                     new NoBibtexFieldChecker(),
+                    new UnicodeNormalFormCCheck(),
                     new BibTeXEntryTypeChecker())
             );
         }